move ProtectedRoute to separate file
